requirements.txt
# Monitoring
snakeviz

# Config system
hydra-core

# Code quality
black

# Logging
wandb
tensorboard
tensorboardX

# Utils
matplotlib 
tqdm

# ML libraries
numpy
gym==0.25
git+https://gitlab-research.centralesupelec.fr/stergios.christodoulidis/text-flappy-bird-gym.git